Create  LinkToAdjMat_v2.0
Takes 1 URL and traverses all the links emanating from the URL to create the adjacency matrix. Traversal is limited to the number of pages specified in the second argument which is MAXPAGES. Returned adjacency matrix will be a square matrix whose max size will be MAXPAGES.If input URL is invalid null will be returned which needs to be handled. At any point of time if there are invalid links it ignores them and proceeds without crashing.